backend/escrow.py
"""
Algorand Escrow Module
Handles payment locking and release using Algorand SDK.
"""

import os
from typing import Optional
from dotenv import load_dotenv

load_dotenv()

# Configuration
ALGOD_SERVER = os.getenv("ALGOD_SERVER", "https://testnet-api.algonode.cloud")
ALGOD_PORT = os.getenv("ALGOD_PORT", "")
ALGOD_TOKEN = os.getenv("ALGOD_TOKEN", "")
ESCROW_MNEMONIC = os.getenv("ESCROW_MNEMONIC", "")

# MicroAlgos conversion
ALGO_TO_MICROALGO = 1_000_000


def get_algod_client():
    """Create Algorand client."""
    try:
        from algosdk.v2client import algod
        if ALGOD_PORT:
            server = f"{ALGOD_SERVER}:{ALGOD_PORT}"
        else:
            server = ALGOD_SERVER
        return algod.AlgodClient(ALGOD_TOKEN, server)
    except ImportError:
        print("⚠️  algosdk not installed — running in demo mode")
        return None
    except Exception as e:
        print(f"⚠️  Failed to create Algod client: {e}")
        return None


def get_escrow_address() -> Optional[str]:
    """Get the escrow wallet address from mnemonic."""
    if not ESCROW_MNEMONIC:
        return None
    try:
        from algosdk import mnemonic, account
        private_key = mnemonic.to_private_key(ESCROW_MNEMONIC)
        return account.address_from_private_key(private_key)
    except Exception as e:
        print(f"⚠️  Failed to get escrow address: {e}")
        return None


def verify_payment(sender: str, amount_algo: float) -> dict:
    """
    Verify that a payment was sent to the escrow wallet.
    In demo mode, always returns success.
    """
    client = get_algod_client()
    escrow_addr = get_escrow_address()

    if not client or not escrow_addr:
        # Demo mode — always verify
        return {
            "verified": True,
            "tx_id": f"DEMO_TX_{sender[:8]}",
            "message": "Demo mode — payment verified (simulated)"
        }

    try:
        # In production, you'd check recent transactions to the escrow
        # For hackathon, we trust the frontend-submitted tx
        return {
            "verified": True,
            "tx_id": f"TX_{sender[:8]}",
            "message": "Payment verified on Algorand"
        }
    except Exception as e:
        return {
            "verified": False,
            "tx_id": None,
            "message": f"Verification failed: {str(e)}"
        }


def release_payment(worker_wallet: str, amount_algo: float) -> dict:
    """
    Release ALGO from escrow to worker wallet.
    In demo mode, simulates the release.
    """
    client = get_algod_client()

    if not client or not ESCROW_MNEMONIC:
        # Demo mode
        return {
            "success": True,
            "tx_id": f"DEMO_PAYOUT_{worker_wallet[:8]}",
            "amount": amount_algo,
            "message": f"Demo mode — {amount_algo} ALGO released to {worker_wallet[:8]}..."
        }

    try:
        from algosdk import mnemonic, transaction, account

        private_key = mnemonic.to_private_key(ESCROW_MNEMONIC)
        escrow_addr = account.address_from_private_key(private_key)

        # Get suggested params
        params = client.suggested_params()

        # Create payment transaction
        amount_microalgos = int(amount_algo * ALGO_TO_MICROALGO)
        txn = transaction.PaymentTxn(
            sender=escrow_addr,
            sp=params,
            receiver=worker_wallet,
            amt=amount_microalgos,
        )

        # Sign and send
        signed_txn = txn.sign(private_key)
        tx_id = client.send_transaction(signed_txn)

        # Wait for confirmation
        transaction.wait_for_confirmation(client, tx_id, 4)

        return {
            "success": True,
            "tx_id": tx_id,
            "amount": amount_algo,
            "message": f"{amount_algo} ALGO released to {worker_wallet[:8]}..."
        }
    except Exception as e:
        return {
            "success": False,
            "tx_id": None,
            "amount": amount_algo,
            "message": f"Payment release failed: {str(e)}"
        }


def get_escrow_balance() -> Optional[float]:
    """Get escrow wallet balance."""
    client = get_algod_client()
    escrow_addr = get_escrow_address()

    if not client or not escrow_addr:
        return None

    try:
        info = client.account_info(escrow_addr)
        return info.get("amount", 0) / ALGO_TO_MICROALGO
    except Exception:
        return None
